% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{BERT}
\alias{BERT}
\title{Load a BERT-like Transformer model}
\usage{
BERT(model_name, path = system.file(package = "golgotha", "models"))
}
\arguments{
\item{model_name}{character string with the name of the model. E.g. 'bert-base-uncased', 'bert-base-multilingual-uncased', 'bert-base-multilingual-cased', 'bert-base-dutch-cased'. Defaults to 'bert-base-multilingual-uncased'.}

\item{path}{path to a directory on disk where the model is stored}
}
\value{
the directory where the model is saved to
}
\description{
Load a BERT-like Transformer model stored on disk
}
\examples{
\dontrun{
bert_download_model("bert-base-multilingual-uncased")
model <- BERT("bert-base-multilingual-uncased")

x <- data.frame(doc_id = c("doc_1", "doc_2"),
                text = c("provide some words to embed", "another sentence of text"),
                stringsAsFactors = FALSE)
predict(model, x, type = "tokenise")
embedding <- predict(model, x, type = "embed-sentence")
dim(embedding)
embedding <- predict(model, x, type = "embed-token")
str(embedding)
}

\dontrun{
model_dir <- file.path(getwd(), "inst", "models")
bert_download_model("bert-base-multilingual-uncased", path = model_dir)
path  <- file.path(getwd(), "inst", "models", "bert-base-multilingual-uncased")
model <- BERT(path = path)
}
}
