% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{bert_download_model}
\alias{bert_download_model}
\title{Download a BERT-like Transformers model}
\usage{
bert_download_model(
  model_name = "bert-base-multilingual-uncased",
  path = system.file(package = "golgotha", "models")
)
}
\arguments{
\item{model_name}{character string with the name of the model. E.g. 'bert-base-uncased', 'bert-base-multilingual-uncased', 'bert-base-multilingual-cased', 'bert-base-dutch-cased'. Defaults to 'bert-base-multilingual-uncased'.}

\item{path}{path to a directory on disk where the model will be downloaded to inside a subfolder \code{model_name}}
}
\value{
invisibly, the directory where the model is saved to
}
\description{
Download a BERT-like Transformers model
}
\examples{
\dontrun{
bert_download_model("bert-base-multilingual-uncased")
bert_download_model("bert-base-dutch-cased")

path <- file.path(getwd(), "inst", "models")
bert_download_model("bert-base-multilingual-uncased", path = path)
bert_download_model("bert-base-dutch-cased", path = path)
}
}
