% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{predict.Transformer}
\alias{predict.Transformer}
\title{Predict alongside a Transformer model}
\usage{
\method{predict}{Transformer}(
  object,
  newdata,
  type = c("embed-sentence", "embed-token", "tokenise"),
  trace = 10,
  ...
)
}
\arguments{
\item{object}{an object of class Transformer as returned by \code{\link{transformer}}}

\item{newdata}{a data.frame with columns doc_id and text indicating the text to embed}

\item{type}{a character string, either 'embed-sentence', 'embed-token', 'tokenise' to get respectively sentence-level embeddings, token-level embeddings or the wordpiece tokens}

\item{trace}{logical indicating to show a trace of the progress. Defaults to showing every 10 annotated embeddings}

\item{...}{other arguments passed on to the methods}
}
\value{
depending on the argument \code{type} the function returns:
\itemize{
\item embed-sentence: A matrix with the embedding of the text, where the doc_id's are in the rownames
\item embed-token: A list of matrices with token-level embeddings, one for each doc_id. The names of the list are identified by the doc_id. Note that depending on the model you will have CLS / SEP tokens at the start/back and the number of rows of the matrix is also limited by the model
\item tokenise: A list of subword (wordpiece) tokens. The names of the list are identified by the doc_id.
\item generate: generate tokens following the provided text sequence
}
}
\description{
Extract features from the Transformer model namely get
\itemize{
\item the embedding of a sentence
\item the embedding of the tokens of the sentence
\item the tokens of a sentence
}
}
\examples{
\donttest{
transformer_download_model("bert-base-multilingual-uncased")
model <- transformer("bert-base-multilingual-uncased")

x <- data.frame(doc_id = c("doc_1", "doc_2"),
                text = c("provide some words to embed", "another sentence of text"),
                stringsAsFactors = FALSE)
predict(model, x, type = "tokenise")
embedding <- predict(model, x, type = "embed-sentence")
dim(embedding)
embedding <- predict(model, x, type = "embed-token")
str(embedding)
}

unlink(file.path(system.file(package = "golgotha", "models"),
       "bert-base-multilingual-uncased"), recursive = TRUE)
}
