% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{unlink_golgotha}
\alias{unlink_golgotha}
\title{Delete all subfolders of the models directory}
\usage{
unlink_golgotha(x = system.file(package = "golgotha", "models"))
}
\arguments{
\item{x}{the path to the folder of the default models directory where the package stores it models}
}
\value{
a character vector of directories which were removed
}
\description{
Delete all subfolders of the models directory
}
\examples{
unlink_golgotha()
}
