% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{transformer}
\alias{transformer}
\title{Load a Transformer model}
\usage{
transformer(
  model_name,
  path = system.file(package = "golgotha", "models"),
  architecture = "BERT"
)
}
\arguments{
\item{model_name}{character string with the name of the model. E.g. 'bert-base-uncased', 'bert-base-multilingual-uncased', 'distilbert-base-cased', 'bert-base-dutch-cased'. Defaults to 'bert-base-multilingual-uncased'.}

\item{path}{path to a directory on disk where the model is stored}

\item{architecture}{character string with the model architecture family name. E.g. 'BERT', 'DistilBERT', 'Roberta',.... Defaults to 'BERT'.}
}
\value{
the directory where the model is saved to
}
\description{
Load a Transformer model stored on disk
}
\examples{
\dontrun{
transformer_download_model("bert-base-multilingual-uncased")
model <- transformer("bert-base-multilingual-uncased")

x <- data.frame(doc_id = c("doc_1", "doc_2"),
                text = c("provide some words to embed", "another sentence of text"),
                stringsAsFactors = FALSE)
predict(model, x, type = "tokenise")
embedding <- predict(model, x, type = "embed-sentence")
dim(embedding)
embedding <- predict(model, x, type = "embed-token")
str(embedding)
}

\dontrun{
model_dir <- file.path(getwd(), "inst", "models")
transformer_download_model("distilbert-base-multilingual-uncased", architecture = "DistilBERT", path = model_dir)
path  <- file.path(getwd(), "inst", "models", "distilbert-base-multilingual-uncased")
model <- transformer(path = path, architecture = "DistilBERT")
}
}
