% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{transformer_download_model}
\alias{transformer_download_model}
\title{Download a Transformers model}
\usage{
transformer_download_model(
  model_name = "bert-base-multilingual-uncased",
  architecture = "BERT",
  path = system.file(package = "golgotha", "models")
)
}
\arguments{
\item{model_name}{character string with the name of the model. E.g. 'bert-base-uncased', 'bert-base-multilingual-uncased', 'bert-base-multilingual-cased', 'bert-base-dutch-cased'. Defaults to 'bert-base-multilingual-uncased'.}

\item{architecture}{character string with the model architecture family name. E.g. 'BERT', 'DistilBERT', 'Roberta',.... Defaults to 'BERT'.}

\item{path}{path to a directory on disk where the model will be downloaded to inside a subfolder \code{model_name}}
}
\value{
invisibly, the directory where the model is saved to
}
\description{
Download a Transformers model
}
\examples{
\dontrun{
transformer_download_model("bert-base-multilingual-uncased")
transformer_download_model(model_name= "distilbert-base-multilingual-uncased", architecture="DistilBERT")

path <- file.path(getwd(), "inst", "models")
transformer_download_model("bert-base-multilingual-uncased", path = path)
transformer_download_model(model_name= "distilbert-base-multilingual-uncased", architecture="DistilBERT", path = path)
}
}
